+++ docs/CHANGE_LOG.md
@@ -6,6 +6,9 @@
 - Added a new binding option called "yearsToHide", which states the years that should be hidden (defaults to []).
 - Added a new binding option called "showLessAndMoreLabels", which states if the "Less" and "More" labels are shown (defaults to true).
 
+#### **Public Functions:**
+- Added a new public function "updateOptions()", which will update the original binding options for an element and refresh it.
+
 #### **Fixes:**
 - Fixed a minor README.md error.
 - Fixed a fault that prevented the "onExport" custom trigger from firing when the public function "export()" is called.